상점 API 동기화 및 보안 강화 작업 완료 (권)
- 상점/재화 API 서버 명세(int, Long, 필드명) 일치화 및 인벤토리 SSOT 동기화 로직 구현
- 미니게임 실행 시 '최근 방문 페이지' 버튼에 정상 기록되도록 MainMenuAction 및 ViewModel 로직 수정
- 서버 주소를 ServerSecrets.kt로 분리하고 .gitignore에 등록하여 보안 강화
- 하단 메뉴 6개 확장 및 미니게임 진입 버튼 연결 완료

+++ LupaPJ/app/src/main/java/com/example/lupapj/data/remote/model/CurrencyEarnRequest.kt
@@ -8,7 +8,7 @@ import com.google.gson.annotations.SerializedName
  */
 data class CurrencyEarnRequest(
     @SerializedName("amount")
-    val amount: Long,
+    val amount: Int, // [수정됨(권)] Long -> Int
     @SerializedName("total")
-    val total: Long // [수정됨(권)] 서버 GoldRequest에 맞춰 amount, total만 유지하고 source 제거됨
+    val total: Long
 )
